packages/chat/requirements.txt
fastapi[standard]
pydantic-settings[yaml]
rich
qdrant-client
openai
httpx
